[#230] :recycle: refactor: diary type 분리
